feat(chp2-l3): a simple udp sender implementation
it is actually fun given that this is me writing UDP since 2022, and i somehow missed the mark of knowing this quirky behaviour of udp; which literally the code equivalent of it doesn't care about established connections, meaning you would get a 'conn' even when there's no connection. like it is bizzaire knowing, given the constant work with tcp, and how the behaviour of no connection, no 'conn' object.